refactor(api): `user-api.ts` in den Ordner `/profile` ausgelagert und Imports angepasst
Refs #232
